Map tab stickers to object types
