--- assets/js/apps/search/components/TabStickers/DatafileTabSticker.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import TabSticker from './TabSticker';
-export const DatafileTabSticker = () => {
-    return (<TabSticker initials="DF"></TabSticker>);
-};
-
-export default DatafileTabSticker;
