--- assets/js/apps/search/components/TabStickers/ProjectTabSticker.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import TabSticker from './TabSticker';
-export const ProjectTabSticker = () => {
-    return (<TabSticker initials="P"></TabSticker>);
-};
-
-
-export default ProjectTabSticker;
